--- nextjs/components/Feed/index.tsx
@@ -1,1 +1,0 @@
-export { Feed } from './Feed';
