--- nextjs/components/MessageCard/index.tsx
@@ -1,1 +1,0 @@
-export { MessageCard } from './MessageCard';
